requirements.txt
beautifulsoup4==4.6.3
Flask==1.0.2
Flask-RESTful==0.3.6
requests==2.19.1
gunicorn==19.7.1
Flask-SQLAlchemy==2.3.2
